requirements.txt
flask
gunicorn
pytest
werkzeug
openai>=1.0.0
python-dotenv
python-pptx
requests